+++ src/IdeHelper.php
@@ -108,6 +108,7 @@ namespace AlibabaCloud;
  * @method static Jarvis\Jarvis jarvis()
  * @method static JarvisPublic\JarvisPublic jarvisPublic()
  * @method static Kms\Kms kms()
+ * @method static Ledgerdb\Ledgerdb ledgerdb()
  * @method static LinkFace\LinkFace linkFace()
  * @method static LinkWAN\LinkWAN linkWAN()
  * @method static Linkedmall\Linkedmall linkedmall()