+++ src/Ledgerdb/Ledgerdb.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace AlibabaCloud\Ledgerdb;
+
+use AlibabaCloud\Client\Resolver\VersionResolver;
+
+/**
+ * @method static V20191122\LedgerdbApiResolver v20191122()
+ */
+class Ledgerdb extends VersionResolver
+{
+}